Creates EJBs to handle email and user registration
Creates simple User entity(to be modified later).
Creates EJB RegisterEJB to handle the registration of an user, as well as the confirmation of the registration.
Creates EJB MailerEJB to handle the sending of the registration email.
Creates test class RegisterEJBTest to test EJB
Creates stub MailerStub for testing purposes.