Adds the dataset and test utility classes for User and Registration
Dataset classes and the companion DBUnit XML datasets have been added to
the tests module to ease the developing of tests with this entities. In
addition, two new Hamcrest matchers for comparing User and Registration
entities have been added.

This commit also completes the Javadoc of the existent test utility
classes and updates the package of the domain entities.