Adds view and controller of index page
Index JSF template has been modified to retrieve the required events from the
EventEJB, with the aid of a new IndexPageController class. There are not yet
tests to check this functionality.

Common functionality of JSF controllers has been also moved to a upper level,
which at the moment has been implemented as an interface called JSFController.