Solves several problems with the tests
The event creation form includes a calendar component for selecting the event
date that was causing test execution blocks. This commit hides this component
by using a little trick (clicking in a previous component and waiting a few
milliseconds).

In addition, the usage of beans.xml and TestPrincipal in
JoinEventControllerTest was causing an unexpected behaviour, as the Principal
didn't returned "anonymous" when no user where logged. This test case also had
several implementation errors and the LoginPage class, on which this test case
depends, where inconsistent with the HTML content of the login.xhtml page.

Finally, the EventsDatset has been refactorized to remove several unused
methods and to add the id to the events.

Despite this changes, there are several elements of this classes that can be
improved.