Adds a TestingMailer to the JSF module
TestingMailer is a stub implementation of the Mailer interface intended
to be used in the tests. This class replaces the DefaultMailer to store
the emails sent instead of effectively sending them.

This email also adds the Email and MailBox classes to the tests module.
This classes can be used to ease the implementation of stub Mailer.
Actually, the JSF's TestingMailer extends the MailBox class.