tsk14 Solves some inconsistencies in the event join action
The current implementation of the event joining action was inconsistent
and caused some unexpected results arbitrarily. The main cause was the
use of forms inside an repeat element in the JSF layer. This
implementation has been changed and, instead of using a form, a link
redirects the join request to a new joinEvent.xhtml page, that processes
the request.