Adds functionalities for anonymous user check to the LoginUserController
This new functionalities make easier to check whether the current user
is an anonymous user or not. Additionally, redirecting functions based
on user anonymity were added to avoid illegal page access. This
functions can be used as "preView" events, to avoid page rendering when
the user can't access a certaing page.