tsk27 Adds API method to list Events sorted by date
REST resource for Events has been created, with a method (GET to event's
resource root) to list events sorted by date in the same way as the index page
does. Some changes have also been done to make it work as expected.

Correct testing of this list method requires changes in events datasets to
validate that date ordering is preserved. It is a pending issue that will be
addressed in following commits.