Adds jsf view and controller to list joined events
Added a view to list the events joined by an user(listJoinedEvent.xhtml).
Added controller with pagination to list joined events(Both links and button pagination).
Updated menu in templateHeader to show the options of a logged in user.
Uncommented a line in templateHeader to allow the use of Jquery