Updates the event search
Updates the visual aspect of the event search functionality to make it more similar to the rest of the web.
Ids were added to the templatesHeader.xhtml file to be used in EventSearchPage and EventSearchForm. This two files were created
but have not been used in an actual test yet.