Task 44 Improvements in Create Event
The calendar component has been fixed by calling jquey.noConflict() in templateLayout.
An event_created paeg has been created to be redirected there when an event is created.
The link to the created events page has been fixed in the dropdown menu.
Refactor: the class LazyEventList has been deleted since it is not used anymore.