tsk58-59 Adds the attribute cancelled to Event
Modifies Event to include the attribute cancelled.
Modifies EventEJB to support the new attribute and includes the new
exception EventIsCancelledException (now an User can't join a cancelled
Event and a cancelled Event will not appear in a search).
Modifies Tests to support the new attribute.
Modifies Datasets to support the new attribute.
Modifies DB (sql) to support the new attribute.