Task 11 Fixes error with certain queries and several improvements more
A message that shows when there isn't any results has been added.
The pagination has been changed to a model where the redirection/pagination is performed via javascript.
CSS has been applied to the no results message.
A modification has been applied to the method count() of eventEJB since the returning value was not the one expected.The affected tests were upgraded and an additional test was added.