tsk59-85 Adds capitals in create/modify Event and fixes search
Modifies rest and jsf persistence.xml to include the Capital 
and CapitalDistances entities.
Modifies createEvent.xhtml and modifyEvent.xhtml to include the place
field.
Modifies CreateEventController and ModifyEventController 
to include the place field.
Modifies EventEJB to include a method to get all the Capitals.
Modifies EventSearchController to hide all the passed events
in the search (without any parameter).
Modifies EventEJBTest to support the new search.