tsk79 Enables user role edition for Administrators
A new field has been added to the user edition view (only visible for Admins)
to allow any administrator to change the current Role of any user.

Another field showing the User's login (read only) has also been added to that
form. The login is the primary key and we cannot update it seamlessly at the
moment, it requires domain-level changes.