tsk97 Adds generation of notifications when an event is cancelled or modified
Modifies EventEJB so that when an event is cancelled or modified, its attendees receive a
notification. It also sends the notification by email to those attendees who have previously
chosen this feature in their own profile.

Modifies testing in the Service and Rest layers so they support Notifications.

Adds new datasets and modifies previous ones to use in the testing classes.

Modifies the order of deletion in cleanup.sql so that it does not leave invalid foreign keys.