tsk93 Adds searching with location capabilities
Modifies EventEJB and EventSearchController so that a location parameter may be added to the search
criteria.

Modifies EventEJB and EventSearchController so that the search functions return events ordered by
physical proximity, date, and number of attendees.

Adds a picklist of Capital cities to templateHeader.xhtml in order to add a location to the search
criteria.

Adds the location GET parameter to the Next and Prev buttons of searchResults.xhtml, to preserve
the location previously chosen.