tsk72 Adds exploring by event category
Modifies EventSearchController and EventEJB so that they support advance searching that takes into
account the state and category of an event.

Adds an Explore button to templateHeader.xhtml to access exploreEvents.xhtml.

Adds exploreEvents.xhtml, which displays a list of all the event categories. The categories are
clickable and lead to advancedSearchResults.xhtml.

Adds advancedSearchResults.xhtml, which enables advanced searching of events by category, state,
and text.

Modifies template.properties and template_es_ES.properties so that they include the string of the
new Explore button at templateHeader.xhtml.